pyproject.toml
[project]
name = "midas"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "databricks-sql-connector>=3.6.0",
    "openai>=1.58.0",
    "pdfplumber>=0.11.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
]

[tool.apx.metadata]
app-name = "midas"
app-slug = "midas"
app-entrypoint = "midas.backend.app:app"
api-prefix = "/api"
metadata-path = "src/midas/_metadata.py"

[tool.apx.ui]
root = "src/midas/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.version]
path = "src/midas/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/midas/__dist__", "src/midas/_metadata.py"]
exclude = ["src/midas/ui"]
